feat: userInputInfoService, userService 구현

+++ src/main/java/com/myZipPlan/server/advice/userInputInfo/dto/UserInputSummaryResponse.java
@@ -3,6 +3,7 @@ package com.myZipPlan.server.advice.userInputInfo.dto;
 import lombok.Builder;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -13,7 +14,7 @@ public class UserInputSummaryResponse {
     private LocalDateTime inquiryDateTime; // 조회일시
     private String dongName;   // 읍명동이름
     private String buildingName;   // 건물명
-    private Long rentalDeposit; // 임차보증금
-    private Long monthlyRent;   // 월세
+    private BigDecimal rentalDeposit; // 임차보증금
+    private BigDecimal monthlyRent;   // 월세
 
 }
